feat: move main into hub and start from cli
