+++ crates/hub/Cargo.toml
@@ -16,7 +16,6 @@ teleport_cli = { path = "../cli" }
 teleport_storage = { path = "../storage" }
 teleport_eth = { path = "../eth" }
 libp2p = { workspace = true }
-clap = { workspace = true }
 prost = { workspace = true }
 tokio = { workspace = true }
 void = { workspace = true }
@@ -32,3 +31,4 @@ thiserror = { workspace = true }
 once_cell = { workspace = true }
 chrono = { workspace = true }
 serde = { workspace = true }
+clap = { workspace = true }
