+++ crates/hub/src/key.rs
@@ -1,19 +1,21 @@
-use crate::key::ed25519::{Keypair, PublicKey, SecretKey};
-use libp2p::identity::ed25519;
+use libp2p::identity::{
+    ed25519::{self, SecretKey},
+    Keypair, PublicKey,
+};
 use teleport_common::config::Config;
 
 pub struct Key {
-    pubkey: PublicKey,
-    privkey: SecretKey,
-    id: Keypair,
+    pub pubkey: PublicKey,
+    pub privkey: SecretKey,
+    pub id: Keypair,
 }
 
 impl Key {
-    fn new(config: Config) -> Self {
+    pub fn new(config: Config) -> Self {
         let secret_key_hex = config.farcaster_priv_key;
         let mut secret_key_bytes = hex::decode(secret_key_hex).expect("Invalid hex string");
-        let privkey = ed25519::SecretKey::try_from_bytes(&mut secret_key_bytes).unwrap();
-        let keypair = ed25519::Keypair::from(privkey);
+        let privkey = SecretKey::try_from_bytes(&mut secret_key_bytes).unwrap();
+        let keypair = ed25519::Keypair::from(privkey.clone());
         let pubkey = keypair.public();
 
         let id_keypair =
@@ -22,7 +24,7 @@ impl Key {
         log::info!("Public Key: {}", hex::encode(pubkey.to_bytes()));
 
         Self {
-            pubkey,
+            pubkey: pubkey.into(),
             privkey,
             id: id_keypair,
         }
